research_agent/requirements.txt
strands-agents[openai]
strands-agents-tools
python-dotenv
mcp[cli]